refactor: move numba utilities to numba connect
