+++ src/awkward/_util.py
@@ -278,36 +278,6 @@ def find_ufunc(behavior, signature):
                 return custom
 
 
-def find_numba_array_typer(layouttype, behavior):
-    behavior = overlay_behavior(behavior)
-    arr = layouttype.parameters.get("__array__")
-    if isinstance(arr, str):
-        typer = behavior.get(("__numba_typer__", arr))
-        if callable(typer):
-            return typer
-    deeprec = layouttype.parameters.get("__record__")
-    if isinstance(deeprec, str):
-        typer = behavior.get(("__numba_typer__", "*", deeprec))
-        if callable(typer):
-            return typer
-    return None
-
-
-def find_numba_array_lower(layouttype, behavior):
-    behavior = overlay_behavior(behavior)
-    arr = layouttype.parameters.get("__array__")
-    if isinstance(arr, str):
-        lower = behavior.get(("__numba_lower__", arr))
-        if callable(lower):
-            return lower
-    deeprec = layouttype.parameters.get("__record__")
-    if isinstance(deeprec, str):
-        lower = behavior.get(("__numba_lower__", "*", deeprec))
-        if callable(lower):
-            return lower
-    return None
-
-
 def find_typestrs(behavior):
     behavior = overlay_behavior(behavior)
     out = {}
@@ -336,107 +306,6 @@ def find_typestr(parameters, typestrs):
             if typestr is not None:
                 return typestr
     return None
-
-
-def find_numba_record_typer(layouttype, behavior):
-    behavior = overlay_behavior(behavior)
-    rec = layouttype.parameters.get("__record__")
-    if isinstance(rec, str):
-        typer = behavior.get(("__numba_typer__", rec))
-        if callable(typer):
-            return typer
-    return None
-
-
-def find_numba_record_lower(layouttype, behavior):
-    behavior = overlay_behavior(behavior)
-    rec = layouttype.parameters.get("__record__")
-    if isinstance(rec, str):
-        lower = behavior.get(("__numba_lower__", rec))
-        if callable(lower):
-            return lower
-    return None
-
-
-def find_numba_attrs(layouttype, behavior):
-    behavior = overlay_behavior(behavior)
-    rec = layouttype.parameters.get("__record__")
-    if isinstance(rec, str):
-        for key, typer in behavior.items():
-            if (
-                isinstance(key, tuple)
-                and len(key) == 3
-                and key[0] == "__numba_typer__"
-                and key[1] == rec
-            ):
-                lower = behavior["__numba_lower__", key[1], key[2]]
-                yield key[2], typer, lower
-
-
-def find_numba_methods(layouttype, behavior):
-    behavior = overlay_behavior(behavior)
-    rec = layouttype.parameters.get("__record__")
-    if isinstance(rec, str):
-        for key, typer in behavior.items():
-            if (
-                isinstance(key, tuple)
-                and len(key) == 4
-                and key[0] == "__numba_typer__"
-                and key[1] == rec
-                and key[3] == ()
-            ):
-                lower = behavior["__numba_lower__", key[1], key[2], ()]
-                yield key[2], typer, lower
-
-
-def find_numba_unaryops(unaryop, left, behavior):
-    behavior = overlay_behavior(behavior)
-    done = False
-
-    if isinstance(left, ak._connect.numba.layout.ContentType):
-        left = left.parameters.get("__record__")
-        if not isinstance(left, str):
-            done = True
-
-    if not done:
-        for key, typer in behavior.items():
-            if (
-                isinstance(key, tuple)
-                and len(key) == 3
-                and key[0] == "__numba_typer__"
-                and key[1] == unaryop
-                and key[2] == left
-            ):
-                lower = behavior["__numba_lower__", key[1], key[2]]
-                yield typer, lower
-
-
-def find_numba_binops(binop, left, right, behavior):
-    behavior = overlay_behavior(behavior)
-    done = False
-
-    if isinstance(left, ak._connect.numba.layout.ContentType):
-        left = left.parameters.get("__record__")
-        if not isinstance(left, str):
-            done = True
-
-    if isinstance(right, ak._connect.numba.layout.ContentType):
-        right = right.parameters.get("__record__")
-        if not isinstance(right, str):
-            done = True
-
-    if not done:
-        for key, typer in behavior.items():
-            if (
-                isinstance(key, tuple)
-                and len(key) == 4
-                and key[0] == "__numba_typer__"
-                and key[1] == left
-                and key[2] == binop
-                and key[3] == right
-            ):
-                lower = behavior["__numba_lower__", key[1], key[2], key[3]]
-                yield typer, lower
 
 
 def behavior_of(*arrays, **kwargs):
